latest

+++ app/models/user.rb
@@ -1,59 +1,74 @@
-class User < ActiveRecord::Base
-  include AvatarInitials
+class User < ApplicationRecord
+  include CustomCounterCache::Model
   include ReadableUnguessableUrls
   include MessageChannel
+  include HasExperiences
+  include HasAvatar
+  include SelfReferencing
+  include NoForbiddenEmails
+  include HasMailer
+  include CustomCounterCache::Model
 
-  AVATAR_KINDS = %w[initials uploaded gravatar]
-  LARGE_IMAGE = 170
-  MED_LARGE_IMAGE = 70
-  MEDIUM_IMAGE = 50
-  SMALL_IMAGE = 30
+  extend HasTokens
+  extend HasDefaults
+
+  extend NoSpam
+  no_spam_for :name
+
+  has_paper_trail only: [:email_newsletter]
+
   MAX_AVATAR_IMAGE_SIZE_CONST = 100.megabytes
+  BOT_EMAILS = {
+    helper_bot: ENV['HELPER_BOT_EMAIL'] || 'contact@loomio.org',
+    demo_bot:   ENV['DEMO_BOT_EMAIL'] || 'contact+demo@loomio.org'
+  }.freeze
 
-  devise :database_authenticatable, :recoverable, :registerable, :rememberable, :trackable, :omniauthable
-  attr_accessor :honeypot
+  devise :database_authenticatable, :recoverable, :registerable, :rememberable
+  attr_accessor :recaptcha
+  attr_accessor :restricted
+  attr_accessor :token
+  attr_accessor :membership_token
 
-  validates :email, presence: true, uniqueness: true, email: true
-  #validates :name, presence: true
-  validates_inclusion_of :uses_markdown, in: [true,false]
+  attr_accessor :legal_accepted
+
+  attr_writer   :has_password
+  attr_accessor :require_valid_signup
+  attr_accessor :require_recaptcha
+
+  before_save :set_legal_accepted_at, if: :legal_accepted
+
+  validates :email, presence: true, email: true, length: {maximum: 200}
+
+  validates :name,               presence: true, if: :require_valid_signup
+  validates :legal_accepted,     presence: true, if: :require_legal_accepted
+  validate  :validate_recaptcha,                 if: :require_recaptcha
 
   has_attached_file :uploaded_avatar,
     styles: {
-              large: "#{User::LARGE_IMAGE}x#{User::LARGE_IMAGE}#",
-              medlarge: "#{User::MED_LARGE_IMAGE}x#{User::MED_LARGE_IMAGE}#",
-              medium: "#{User::MEDIUM_IMAGE}x#{User::MEDIUM_IMAGE}#",
-              small: "#{User::SMALL_IMAGE}x#{User::SMALL_IMAGE}#",
-            }
+      small:  "#{AVATAR_SIZES[:small]}x#{AVATAR_SIZES[:small]}#",
+      medium: "#{AVATAR_SIZES[:medium]}x#{AVATAR_SIZES[:medium]}#",
+      large:  "#{AVATAR_SIZES[:large]}x#{AVATAR_SIZES[:large]}#",
+    }
+
   validates_attachment :uploaded_avatar,
-    size: { in: 0..User::MAX_AVATAR_IMAGE_SIZE_CONST.kilobytes },
-    content_type: { content_type: /\Aimage/ },
-    file_name: { matches: [/png\Z/i, /jpe?g\Z/i, /gif\Z/i] }
+    size: { in: 0..MAX_AVATAR_IMAGE_SIZE_CONST.kilobytes },
+    content_type: { content_type: /\Aimage/ }
 
-  validates_inclusion_of :avatar_kind, in: AVATAR_KINDS
-
-  validates_uniqueness_of :username
+  validates_uniqueness_of :email, conditions: -> { where(email_verified: true) }, if: :email_verified?
+  validates_uniqueness_of :username, if: :email_verified
+  before_validation :generate_username, if: :email_verified
   validates_length_of :username, maximum: 30
-  validates_format_of :username, with: /\A[a-z0-9]*\z/, message: I18n.t(:'error.username_must_be_alphanumeric')
+  validates_length_of :short_bio, maximum: 500
+  validates_format_of :username, with: /\A[a-z0-9]*\z/, message: I18n.t(:'user.error.username_must_be_alphanumeric')
+  validates_confirmation_of :password, if: :password_required?
 
   validates_length_of :password, minimum: 8, allow_nil: true
   validates :password, nontrivial_password: true, allow_nil: true
 
-  include Gravtastic
-  gravtastic rating: :pg, default: :none
-  before_create :set_interface
-
-
-  has_many :contacts, dependent: :destroy
   has_many :admin_memberships,
            -> { where('memberships.admin = ? AND memberships.is_suspended = ?', true, false) },
            class_name: 'Membership',
            dependent: :destroy
-
-  has_many :adminable_groups,
-           -> { where( archived_at: nil) },
-           through: :admin_memberships,
-           class_name: 'Group',
-           source: :group
 
   has_many :memberships,
            -> { where(is_suspended: false, archived_at: nil) },
@@ -62,6 +77,22 @@ class User < ActiveRecord::Base
   has_many :archived_memberships,
            -> { where('archived_at IS NOT NULL') },
            class_name: 'Membership'
+
+  has_many :invited_memberships,
+           class_name: 'Membership',
+           foreign_key: :inviter_id
+
+  has_many :formal_groups,
+           -> { where(type: "FormalGroup") },
+           through: :memberships,
+           class_name: 'FormalGroup',
+           source: :group
+
+  has_many :adminable_groups,
+           -> { where(archived_at: nil) },
+           through: :admin_memberships,
+           class_name: 'Group',
+           source: :group
 
   has_many :membership_requests,
            foreign_key: 'requestor_id',
@@ -79,98 +110,166 @@ class User < ActiveRecord::Base
            foreign_key: 'author_id',
            dependent: :destroy
 
-  has_many :motions,
-           through: :discussions
+  has_many :polls, foreign_key: :author_id
 
-  has_many :authored_motions,
-           class_name: 'Motion',
-           foreign_key: 'author_id',
-           dependent: :destroy
+  has_many :identities, class_name: "Identities::Base", dependent: :destroy
 
-  has_many :votes, dependent: :destroy
-  has_many :comment_votes, dependent: :destroy
+  has_many :reactions, dependent: :destroy
+  has_many :stances, foreign_key: :participant_id, dependent: :destroy
+  has_many :participated_polls, through: :stances, source: :poll
+  has_many :group_polls, through: :groups, source: :polls
 
   has_many :discussion_readers, dependent: :destroy
-  has_many :motion_readers, dependent: :destroy
-  has_many :omniauth_identities, dependent: :destroy
-
 
   has_many :notifications, dependent: :destroy
   has_many :comments, dependent: :destroy
-  has_many :attachments, dependent: :destroy
+  has_many :documents, foreign_key: :author_id, dependent: :destroy
   has_many :drafts, dependent: :destroy
+  has_many :login_tokens, dependent: :destroy
 
   has_one :deactivation_response,
           class_name: 'UserDeactivationResponse',
           dependent: :destroy
 
-  before_validation :generate_username
-  before_save :set_avatar_initials,
-              :ensure_unsubscribe_token,
-              :ensure_email_api_key
+  before_save :set_avatar_initials
+  initialized_with_token :unsubscribe_token,        -> { Devise.friendly_token }
+  initialized_with_token :email_api_key,            -> { SecureRandom.hex(16) }
+  initialized_with_default :email_on_participation, -> { !ENV['EMAIL_ON_PARTICIPATION_DEFAULT_FALSE'] }
 
-  before_create :set_default_avatar_kind
+
+  enum default_membership_volume: [:mute, :quiet, :normal, :loud]
 
   scope :active, -> { where(deactivated_at: nil) }
   scope :inactive, -> { where("deactivated_at IS NOT NULL") }
-  scope :email_missed_yesterday, -> { active.where(email_missed_yesterday: true) }
+  scope :email_catch_up, -> { active.verified.where(email_catch_up: true) }
   scope :sorted_by_name, -> { order("lower(name)") }
   scope :admins, -> { where(is_admin: true) }
   scope :coordinators, -> { joins(:memberships).where('memberships.admin = ?', true).group('users.id') }
-
-  # move to ThreadMailerQuery
+  scope :verified, -> { where(email_verified: true) }
+  scope :unverified, -> { where(email_verified: false) }
+  scope :verified_first, -> { order(email_verified: :desc) }
+  scope :search_for, -> (q) { where("users.name ilike :first OR users.name ilike :other OR users.username ilike :first", first: "#{q}%", other:  "% #{q}%") }
+  scope :visible_by, -> (user) { distinct.active.verified.joins(:memberships).where("memberships.group_id": user.group_ids).where.not(id: user.id) }
+  scope :mention_search, -> (user, model, query) do
+    # allow mentioning of anyone in the organisation
+    group_ids = (model.group.parent_or_self.id_and_subgroup_ids + [model.guest_group.id]).compact.uniq
+    distinct.active.verified.
+      search_for(query).
+      joins(:memberships).
+      where("memberships.group_id": group_ids).
+      where.not(id: user.id).
+      order("users.name")
+  end
   scope :email_when_proposal_closing_soon, -> { active.where(email_when_proposal_closing_soon: true) }
 
   scope :email_proposal_closing_soon_for, -> (group) {
-    active.
-    joins(:memberships).
-    where('memberships.group_id = ?', group.id).
-    where('users.email_when_proposal_closing_soon = ?', true)
+     email_when_proposal_closing_soon
+    .joins(:memberships)
+    .where('memberships.group_id': group.id)
   }
 
-  scope :without, -> (users) {
-    users = Array(users).compact
-
-    if users.size > 0
-      where('users.id NOT IN (?)', users)
-    else
-      all
-    end
+  scope :joins_readers, ->(model) {
+    joins("LEFT OUTER JOIN discussion_readers dr ON (dr.user_id = users.id AND dr.discussion_id = #{model.discussion_id.to_i})")
   }
 
-  def self.email_taken?(email)
-    User.find_by_email(email).present?
+  scope :joins_formal_memberships, ->(model) {
+     joins("LEFT OUTER JOIN memberships fm ON (fm.user_id = users.id AND fm.group_id = #{model.group_id.to_i})")
+    .where('fm.archived_at': nil)
+  }
+
+  scope :joins_guest_memberships, ->(model) {
+     joins("LEFT OUTER JOIN memberships gm ON (gm.user_id = users.id AND gm.group_id = #{model.guest_group_id.to_i})")
+    .where('gm.archived_at': nil)
+  }
+
+  # This is a double-nested join select raw sql statement, eek!
+  # But, it's not soo complicated. Here's what's going on:
+  # Join 1: Grab all instances of a user receiving an announcement from the given model, based on the model's announcement ids
+  # Join 2: Group those instances, taking the most recent instance's created_at as the last_notified_at timestamp
+  #
+  # then, we join that timestamp to the current user query, available in the last_notified_at column
+  # scope :with_last_notified_at, ->(model) {
+  #   select('users.*, last_notified_at').joins(<<~SQL)
+  #     -- join #2
+  #     LEFT JOIN (
+  #       SELECT users.id as user_id, max(notified.created_at) as last_notified_at
+  #       FROM users
+  #       -- join #1
+  #       LEFT JOIN (
+  #         SELECT user_ids, created_at
+  #         FROM   announcees
+  #         WHERE  announcees.announcement_id IN (#{model.announcement_ids.join(',').presence || '-1'})
+  #       ) notified ON notified.user_ids ? users.id::varchar
+  #       GROUP BY users.id
+  #     ) announcements ON announcements.user_id = users.id
+  #   SQL
+  # }
+  #
+
+  def set_legal_accepted_at
+    self.legal_accepted_at = Time.now
   end
 
-  def user_id
-    id
+  def require_legal_accepted
+    self.require_valid_signup && ENV['TERMS_URL']
+  end
+
+  def self.email_status_for(email)
+    verified_first.find_by(email: email)&.email_status || :unused
+  end
+
+  def self.find_for_database_authentication(warden_conditions)
+    super(warden_conditions.merge(email_verified: true))
+  end
+
+  define_counter_cache(:memberships_count) {|user| user.memberships.formal.count }
+
+  def associate_with_identity(identity)
+    if existing = identities.find_by(user: self, uid: identity.uid, identity_type: identity.identity_type)
+      existing.update(access_token: identity.access_token)
+    else
+      identities.push(identity)
+      identity.assign_logo! if avatar_kind == 'initials'
+    end
+    self
+  end
+
+  def identity_for(type)
+    identities.find_by(identity_type: type)
+  end
+
+  def pending_invitation_limit
+    ENV.fetch('MAX_PENDING_INVITATIONS', 100).to_i +
+    self.invited_memberships.accepted.count        -
+    self.invited_memberships.pending.count
+  end
+
+  def verified_or_self
+    self.class.verified.find_by(email: email) || self
+  end
+
+  def first_name
+    name.split(' ').first
+  end
+
+  def last_name
+    name.split(' ').drop(1).join(' ')
+  end
+
+  def remember_me
+    true
   end
 
   def is_logged_in?
     true
   end
 
-  def is_logged_out?
-    !is_logged_in?
+  def has_password
+    self.encrypted_password.present?
   end
 
-  def top_level_groups
-    parents = groups.parents_only.order(:name).includes(:children)
-    orphans = groups.where('parent_id not in (?)', parents.map(&:id))
-    (parents.to_a + orphans.to_a).sort{|a, b| a.full_name <=> b.full_name }
-  end
-
-  def inbox_groups
-    groups.where('memberships.inbox_position is not null').order('memberships.inbox_position')
-  end
-
-  def groups_discussions_can_be_started_in
-    (groups.where(members_can_start_discussions: true) | adminable_groups).
-     sort{|a,b| a.full_name <=> b.full_name}
-  end
-
-  def first_name
-    name.split(' ').first
+  def email_status
+    if deactivated_at.present? then :inactive else :active end
   end
 
   def name_and_email
@@ -179,66 +278,42 @@ class User < ActiveRecord::Base
 
   # Provide can? and cannot? as methods for checking permissions
   def ability
-    @ability ||= Ability.new(self)
+    @ability ||= Ability::Base.new(self)
   end
 
   delegate :can?, :cannot?, :to => :ability
 
-  def voting_motions
-    motions.voting
-  end
-
-  def closed_motions
-    motions.closed
-  end
-
-  def is_group_admin?(group=nil)
-    if group.present?
-      admin_memberships.where(group_id: group.id).any?
-    else
-      admin_memberships.any?
-    end
-  end
-
   def is_member_of?(group)
-    memberships.where(group_id: group.id).any?
+    !!memberships.find_by(group_id: group&.id)
   end
 
-  def time_zone_city
-    TimeZoneToCity.convert time_zone
+  def is_admin_of?(group)
+    !!memberships.find_by(group_id: group&.id, admin: true)
+  end
+
+  def first_name
+    self.name.to_s.split(' ').first
   end
 
   def time_zone
     self[:time_zone] || 'UTC'
   end
 
-  def is_parent_group_member?(group)
-    memberships.for_group(group.parent).exists? if group.parent
+  def self.helper_bot
+    verified.find_by(email: BOT_EMAILS[:helper_bot]) ||
+    create!(email: BOT_EMAILS[:helper_bot],
+            name: 'Loomio Helper Bot',
+            password: SecureRandom.hex(20),
+            email_verified: true,
+            avatar_kind: :gravatar)
   end
 
-  def group_membership(group)
-    memberships.for_group(group).first
-  end
-
-  def self.loomio_helper_bot(password: nil)
-    where(email: 'contact@loom.io').first ||
-    create!(email: 'contact@loom.io', name: 'Loomio Helper Bot', password: password || SecureRandom.hex)
-  end
-
-  def self.helper_bots
-    where(email: ['contact@loomio.org', 'contact@loom.io'])
-  end
-
-  def self.find_by_email(email)
-    User.where('lower(email) = ?', email.downcase).first
-  end
-
-  def subgroups
-    groups.where("parent_id IS NOT NULL")
-  end
-
-  def parent_groups
-    groups.where("parent_id IS NULL").order("LOWER(name)")
+  def self.demo_bot
+    verified.find_by(email: BOT_EMAILS[:helper_bot]) ||
+    create!(email: BOT_EMAILS[:demo_bot],
+            name: 'Loomio Demo bot',
+            email_verified: true,
+            avatar_kind: :gravatar)
   end
 
   def name
@@ -250,19 +325,12 @@ class User < ActiveRecord::Base
   end
 
   def deactivate!
+    return if self.deactivated_at
     former_group_ids = group_ids
     update_attributes(deactivated_at: Time.now, avatar_kind: "initials")
     memberships.update_all(archived_at: Time.now)
     Group.where(id: former_group_ids).map(&:update_memberships_count)
     membership_requests.where("responded_at IS NULL").destroy_all
-  end
-
-  def deactivated?
-    deactivated_at.present?
-  end
-
-  def active?
-    deactivated_at.nil?
   end
 
   def reactivate!
@@ -275,98 +343,34 @@ class User < ActiveRecord::Base
     super && !deactivated_at
   end
 
-  def inactive_message
-    I18n.t(:inactive_html, path_to_contact: '/contact').html_safe
-  end
-
-  def avatar_url(size=nil)
-    size = size ? size.to_sym : :medium
-    case size
-    when :small
-      pixels = User::SMALL_IMAGE
-    when :medium
-      pixels = User::MEDIUM_IMAGE
-    when :"med-large"
-      pixels = User::MED_LARGE_IMAGE
-    when :large
-      pixels = User::LARGE_IMAGE
-    else
-      pixels = User::SMALL_IMAGE
-    end
-
-    if avatar_kind == "gravatar"
-      gravatar_url(:size => pixels)
-    else
-      uploaded_avatar.url(size)
-    end
-  end
-
   def locale
-    selected_locale || detected_locale || I18n.default_locale
+    selected_locale || detected_locale || I18n.locale
   end
 
-  def using_initials?
-    avatar_kind == "initials"
-  end
-
-  def has_uploaded_image?
-    uploaded_avatar.url(:medium) != '/uploaded_avatars/medium/missing.png'
-  end
-
-  def has_gravatar?(options = {})
-    return false if Rails.env.test?
-    hash = Digest::MD5.hexdigest(email.to_s.downcase)
-    options = { :rating => 'x', :timeout => 2 }.merge(options)
-    http = Net::HTTP.new('www.gravatar.com', 80)
-    http.read_timeout = options[:timeout]
-    response = http.request_head("/avatar/#{hash}?rating=#{options[:rating]}&default=http://gravatar.com/avatar")
-    response.code != '302'
-  rescue StandardError, Timeout::Error
-    false  # Don't show "gravatar" if the service is down or slow
+  def update_detected_locale(locale)
+    self.update_attribute(:detected_locale, locale) if self.detected_locale&.to_sym != locale.to_sym
   end
 
   def generate_username
-    self.username ||= UsernameGenerator.new(self).generate
+    self.username ||= ::UsernameGenerator.new(self).generate
   end
 
-  def in_same_group_as?(other_user)
-    (group_ids & other_user.group_ids).present?
+  def send_devise_notification(notification, *args)
+    I18n.with_locale(locale) { devise_mailer.send(notification, self, *args).deliver_now }
   end
 
-  def show_start_group_button?
-    !groups.cannot_start_parent_group.any?
-  end
+  protected
 
-  def is_organisation_coordinator?
-    adminable_groups.parents_only.any?
+  def password_required?
+    !password.nil? || !password_confirmation.nil?
   end
 
   private
-  def set_interface
-    if ENV['LOOMIO_NEW_USERS_ON_BETA']
-      self.angular_ui_enabled = false
-    end
-    true
-  end
 
-  def set_default_avatar_kind
-    if has_gravatar?
-      self.avatar_kind = "gravatar"
-    end
-  end
-
-  def ensure_email_api_key
-    self.email_api_key ||= SecureRandom.hex(16)
-  end
-
-  def ensure_unsubscribe_token
-    if unsubscribe_token.blank?
-      found = false
-      while not found
-        token = Devise.friendly_token
-        found = true unless self.class.where(:unsubscribe_token => token).exists?
-      end
-      self.unsubscribe_token = token
-    end
+  def validate_recaptcha
+    return unless ENV['RECAPTCHA_APP_KEY']
+    return if self.persisted?
+    return if Clients::Recaptcha.instance.validate(self.recaptcha)
+    self.errors.add(:recaptcha, I18n.t(:"user.error.recaptcha"))
   end
 end
